requirements.txt
argparse==1.2.1
django-countries==3.3
mock==1.0.1
nose==1.3.3
